backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
authlib==1.3.1
Flask-Session==0.6.0
redis==5.0.8
python-dotenv==1.0.1
flask-cors==4.0.1
requests==2.32.3
passlib[bcrypt]==1.7.4
psycopg2-binary
argon2-cffi==23.1.0
supabase==2.22.0
supabase-auth==2.22.0
supabase-functions==2.22.0

